step2 fill Menu class
step1 Main클래스에서 구현한 input 메소드 2개를 Menu클래스로 수정
