+++ src/com/company/Main.java
@@ -1,8 +1,6 @@
 package com.company;
 
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
 import java.util.Arrays;
 import java.util.Random;
 
@@ -10,27 +8,12 @@ public class Main {
 
     public static void main(String[] args) throws IOException {
         Main m = new Main();
-        m.startRacing();
-        int monsterCount = m.inputMonsterCount();
-        int attemptCount = m.inputAttemptCount();
+
+        Menu menu = new Menu();
+        int monsterCount = menu.inputMonsterCount();
+        int attemptCount = menu.inputAttemptCount();
         int[] result = m.runRacing(monsterCount, attemptCount);
         m.showResult(result);
-    }
-
-    public void startRacing() {
-        System.out.println("<스릴만점 건전한 몬스터 경주>");
-    }
-
-    public int inputMonsterCount() throws IOException {
-        System.out.println("몬스터는 모두 몇 마리인가요?");
-        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        return Integer.parseInt(br.readLine());
-    }
-
-    public int inputAttemptCount() throws IOException {
-        System.out.println("시도할 회수는 몇 회 인가요?");
-        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        return Integer.parseInt(br.readLine());
     }
 
     public int[] runRacing(int monsterCount, int attemptCount) {
